combine startup and permissions windows
